ci: add firebase configuration
This should allow GitHub actions to deploy the app automatically.

Yes, this feels very wrong to expose API keys like this, but it is
just fine to do this with Firebase API keys. It is actually intended
since env vars can't be used in the browser build. These are also
already exposed in the published app anyways if someone was to dig
through the source.