refactor: switch to Composition API
The class-based API from Vue 2 is not being moved forward to Vue 3, so
vue-class-component and vue-property-decorator are no longer able to be
used. I had to choose between the Options API and Composition API to
move this project to Vue 3, so I chose the Composition API as a learning
exercise.