refactor: move database and session logic to pinia
The database store is intended to be the single place where changes are
made to the database, so it is more clear when something is updating the
local state vs. storing a change in the database.

The database store also takes over managing reactive refs to the
database, which was previously handled by a custom FirebaseManager class.